+++ src/utils.ts
@@ -585,6 +585,38 @@ export function commitmentConfig(commitment: Commitment): ConfirmOptions {
     commitment,
   };
 }
+export async function simulateTransaction(
+  provider: anchor.AnchorProvider,
+  tx: Transaction
+) {
+  let response: any;
+  try {
+    response = await provider.simulate(tx);
+  } catch (err) {
+    let parsedErr = parseError(err);
+    throw parsedErr;
+  }
+
+  if (response === undefined) {
+    throw new Error("Unable to simulate transaction");
+  }
+  const logs = response.logs;
+  if (!logs) {
+    throw new Error("Simulated logs not found");
+  }
+
+  let parser = new anchor.EventParser(
+    Exchange.programId,
+    Exchange.program.coder
+  );
+
+  let events = [];
+  parser.parseLogs(response.logs, (event) => {
+    events.push(event);
+  });
+
+  return { events, raw: logs };
+}
 
 export async function processTransaction(
   provider: anchor.AnchorProvider,
@@ -613,29 +645,41 @@ export async function processTransaction(
     );
     return txSig;
   } catch (err) {
-    const anchorError = anchor.AnchorError.parse(err.logs);
-    if (anchorError) {
-      // Parse Anchor error into another type such that it's consistent.
-      throw NativeAnchorError.parse(anchorError);
-    }
-
-    const programError = anchor.ProgramError.parse(err, idlErrors);
-    if (programError) {
-      throw programError;
-    }
-
-    let customErr = parseCustomError(err);
-    if (customErr != null) {
-      throw customErr;
-    }
-
-    let nativeErr = NativeError.parse(err);
-    if (nativeErr != null) {
-      throw nativeErr;
-    }
-
-    throw err;
+    let parsedErr = parseError(err);
+    throw parsedErr;
   }
+}
+
+export function parseError(err: any) {
+  const anchorError = anchor.AnchorError.parse(err.logs);
+  if (anchorError) {
+    // Parse Anchor error into another type such that it's consistent.
+    return NativeAnchorError.parse(anchorError);
+  }
+
+  const programError = anchor.ProgramError.parse(err, idlErrors);
+  if (programError) {
+    return programError;
+  }
+
+  let customErr = parseCustomError(err);
+  if (customErr != null) {
+    return customErr;
+  }
+
+  let nativeErr = NativeError.parse(err);
+  if (nativeErr != null) {
+    return nativeErr;
+  }
+
+  if (err.simulationResponse) {
+    let simulatedError = anchor.AnchorError.parse(err.simulationResponse.logs);
+    if (simulatedError) {
+      return NativeAnchorError.parse(simulatedError);
+    }
+  }
+
+  return err;
 }
 
 const uint64 = (property = "uint64") => {